% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_PhysChemProps.R
\name{AA_PhysChemProps}
\alias{AA_PhysChemProps}
\title{Obtain the physico-chemical properties from a peptide sequence}
\usage{
AA_PhysChemProps(AAfreq)
}
\arguments{
\item{AAfreq}{The relative frequency of amino acids in a peptide sequence}
}
\value{
A data.frame object containing the properties of the amino acid sequence (columns) for the sequences submitted (rows)
}
\description{
Computes different properties for a peptide sequence from the amino acid properties and the amino acid frequency of the sequence.
}
\references{
Adapted from package 'seqinr'.
}
\seealso{
\code{\link{seqinr::AAstat}}
}

