% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AAindexData}
\alias{AAindexData}
\title{List of 544 physicochemical and biological properties for the 20 amino-acids}
\format{A data.frame with 544 rows and 26 columns
Each row represents a property
Columns 1 to 6 give information about the property "AccNo"      "PropDesc"   "LITDBNo"    "Author"     "ArtTitle"   "JournalRef"
Columns 7 to 26 give the property's values for the 20 amino acids}
\source{
http://www.genome.jp/aaindex
}
\usage{
AAindexData
}
\description{
Data were imported from seqinr package.  Refer to the aaindex in seqinr for more details. See the reference section to cite this database in a publication.
}
\references{
From the original aaindex documentation:
 Please cite the following references when making use of the database:

 Kawashima, S. and Kanehisa, M. (2000) AAindex: amino acid index database. Nucleic Acids Res., 28:374.
Tomii, K. and Kanehisa, M. (1996) Analysis of amino acid indices and mutation matrices for sequence comparison and structure prediction of proteins. Protein Eng., 9:27-36.
Nakai, K., Kidera, A., and Kanehisa, M. (1988) Cluster analysis of amino acid indices for prediction of protein structure and function. Protein Eng. 2:93-100.
}
\keyword{datasets}

