% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AAproperty}
\alias{AAproperty}
\title{Amino acid properties
list of amino acid with properties tiny, small, aliphatic, aromatic, non.polar, polar, charged, basic and acidic}
\format{An object of class \code{list} of length 9.}
\usage{
AAproperty
}
\description{
Amino acid properties
list of amino acid with properties tiny, small, aliphatic, aromatic, non.polar, polar, charged, basic and acidic
}
\seealso{
seqinr package
}
\keyword{datasets}
