% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckFeatures.R
\name{CheckFeatures}
\alias{CheckFeatures}
\title{Check valididity of feature table}
\usage{
CheckFeatures(Features)
}
\arguments{
\item{Features}{A data.frame containing the features}
}
\value{
FALSE if found a NA value otherwise TRUE
}
\description{
Check valididity of feature table
}
\author{
Carlus Deneke
}
