% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.TrainingDataSet.R
\name{Create.TrainingDataSet}
\alias{Create.TrainingDataSet}
\title{Create a new training data set}
\usage{
Create.TrainingDataSet(Path2Files = NULL, pattern = "Features",
  OSlabels = NULL, savePath = file.path(Path2Files, "TrainingData"),
  CompressOption = T)
}
\arguments{
\item{Path2Files}{A complete path to the read folder (i.e. it contains a subfolder 'Features')}

\item{pattern}{A string providing a distinct search pattern for all feature files (default = "Features")}

\item{OSlabels}{A vector containing either 'HP' or 'NP' together with the name attributes pointing to the Organism identifier (Bioproject ID)}

\item{savePath}{The name of the newly created training data set folder}

\item{CompressOption}{Do you want to compress the saved files (takes time but saves disk space)?}
}
\value{
Returns True if completed. Feature and label files are saved in specified folder (see details)
#' @examples
\dontrun{
Create.TrainingDataSet (Path2Files = NULL, OSlabels = NULL, savePath = NULL,CompressOption = T)
}
}
\description{
Create a new training data set from a set of feature files. It concatenates all feature file into a data.table object. Read labels are assigned from the corresponding organism labels. Data are saved in specified folders.
}
\details{
Saves 3 files in specified folder "TrainingFolder".
FeatureTable.rds containing the features in a data.frame object
FeatureRowDescription.rds containing the read description and allows identification of Organism, Chromosome and read
ReadLabel_OS.rds containing a label for each read based on the organisms label

This function uses the package data.table for efficiently joining many large data.frames. For even larger files, it is recommended to use linux command line tools
}
\author{
Carlus Deneke
}
\seealso{
Other TrainingFunctions: \code{\link{Run.Training}},
  \code{\link{SelectFeatureSubset}}
}

