% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFeaturesFromReads.R
\name{CreateFeaturesFromReads}
\alias{CreateFeaturesFromReads}
\title{Create features from reads}
\usage{
CreateFeaturesFromReads(Reads, NT_kmax = 4, SymmetricFeatures = T,
  Do.NTMotifs = F, NTMotifs = NULL, AllowedMismatches = 0,
  bothStrands = T, Do.SpacedWords = T, k.spaced = 4, l.spaced = 6,
  SingleSpacerPattern = 0, combineSpacerPatterns = T,
  Do.PeptideFeatures = T, Do.MonoPep = T, Do.DiPep = T, Do.AAprops = T,
  Do.AAindex = T, AAindex_Selection = NULL, Do.UCO = T,
  Do.PepPatterns = F, Patterns = NULL, SearchPatternsSixFrame = F,
  AllowedPeptideMismatches = 0, AggregatePatterns = F)
}
\arguments{
\item{Reads}{A DNAStringSet object}

\item{Do.NTMotifs}{Should features based on specified (genomic) sequence motifs (see NTMotifs) be computed}

\item{NTMotifs}{Vector of characters containing the sequence motifs}

\item{AllowedMismatches}{How many Mismatches are allowed for the motif search}

\item{bothStrands}{Should the motifs be search on both strands?}

\item{Do.SpacedWords}{Should the occurrence of spaced words be computed?}

\item{k.spaced}{Specify the word length}

\item{l.spaced}{Specify the word length including spacers}

\item{SingleSpacerPattern}{Integer value to pick a single spacer pattern of all possible patterns}

\item{combineSpacerPatterns}{combine different patterns: combine column wise: add}

\item{Do.PeptideFeatures}{Should the DNA sequence be translated and peptide features be computed?}

\item{Do.DiPep}{Should the monopeptide frequency be computed?}

\item{Do.AAprops}{Should the amino acid properties be computed?}

\item{Do.AAindex}{Should properties based on the AAindex stats be computed?}

\item{AAindex_Selection}{Which AAindex properties should be computed: Vector of accession numbers}

\item{Do.UCO}{Should the codon usage frequency be computed?}

\item{Do.PepPatterns}{Should peptide pattern counts be computed?}

\item{Patterns}{A vector of charaters for the peptide pattern search}

\item{SearchPatternsSixFrame}{Should the patterns be searched in all 6 frames?}

\item{AggregatePatterns}{Should the indidual patterns be aggregated into a single column?}

\item{Do.DiPep}{Should the dipeptide frequency be computed?}
}
\value{
A data.frame containing the different features as columns. Each row repesents a DNA-string
}
\description{
This function extracts all kind of different features from a set of DNA sequences
}
\details{
Make Stats from DNAString object, based on nucleotide sequence itself and the best translation into an amino acid sequence
}
\examples{
data(ReadData)
CreateFeaturesFromReads (Reads = ReadData)
}
\author{
Carlus Deneke
}

