% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.AAindexSelection.R
\name{default.AAindexSelection}
\alias{default.AAindexSelection}
\title{Print the default selection of aaindices and their description}
\usage{
default.AAindexSelection()
}
\value{
Prints the default selection of aaindices and their description in data.frame format
}
\description{
Print the default selection of aaindices and their description
}
\author{
Carlus Deneke
}
