% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enumerate.SpacerPatterns.R
\name{Enumerate.SpacerPatterns}
\alias{Enumerate.SpacerPatterns}
\title{Enumerate all possible patterns of spacers}
\usage{
Enumerate.SpacerPatterns(k, l)
}
\arguments{
\item{k}{Length of contingous word}

\item{l}{Length of spaced word}
}
\value{
Returns a data.frame of all possible patterns, row-by-row
}
\description{
Enumerate all possible patterns of spacers
}
\author{
Carlus Deneke
}
\seealso{
\link{CreateFeaturesFromReads}

Other SpacedWordFeatures: \code{\link{Count.SpacedWords}}
}

