% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPatternsInAAseq_wMismatches.R
\name{FindPatternsInAAseq_wMismatches}
\alias{FindPatternsInAAseq_wMismatches}
\title{Find patterns of amino acids}
\usage{
FindPatternsInAAseq_wMismatches(Seqs, Patterns, AllowedMismatches = 0,
  aggregate = F, Prefix = "AAPattern")
}
\arguments{
\item{Patterns}{Vector of patterns}

\item{aggregate}{Should the individual pattern matches be aggregated into a single column?}

\item{Prefix}{How should the columns of pattern matches be called}

\item{Seq}{AAStringSet object}
}
\value{
A data.frame containing the matched patterns (columns)
}
\description{
For a list of sequences, search the occurences of specified patterns
}
\author{
Carlus Deneke
}
