% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forest}
\alias{forest}
\title{Example random forest}
\format{An object of class \code{ranger} of length 12.}
\usage{
forest
}
\description{
Example random forest
}
\details{
Provided is a small random forest trained on a very small datasets of pathogens and non-pathogens.
To be used for illustrative examples / tests only
}
\keyword{datasets}
