% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenomicMotifs}
\alias{GenomicMotifs}
\title{Genomic Motifs
100 most frequent 8-mers in bacterial genomes}
\format{An object of class \code{character} of length 100.}
\usage{
GenomicMotifs
}
\description{
Genomic Motifs
100 most frequent 8-mers in bacterial genomes
}
\keyword{datasets}
