% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAAindexStats.R
\name{GetAAindexStats}
\alias{GetAAindexStats}
\title{Get the AAindex property of a peptide sequence}
\usage{
GetAAindexStats(AAfreq, AAindex_Selection)
}
\arguments{
\item{AAfreq}{The relative frequency of amino acids in the peptide sequence obtained from function \code{\link{GetMonoPeptideFrequency}}}

\item{Vector}{of accession numbers to select a subset of the 544 amino acid indices}
}
\value{
A data.frame containing the desired amino acid index stats (columns) for each sequence submitted (rows)
}
\description{
For a peptide sequence, obtain properties associated to its amino acid content
}
\author{
Carlus Deneke
}
\references{
Adapted from aaindex in seqinr

Please cite the following references when making use of the database:

Kawashima, S. and Kanehisa, M. (2000) AAindex: amino acid index database. Nucleic Acids Res., 28:374.
}

