% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBestTranslatedSequence.R
\name{GetBestTranslatedSequence}
\alias{GetBestTranslatedSequence}
\title{Translate nucleotide sequence into a amino acid sequence}
\usage{
GetBestTranslatedSequence(Reads)
}
\arguments{
\item{Reads}{A DNAStringSet Object}
}
\value{
Returns a list containing 'BestSeq' (Peptide sequence with the lowest number of stop codons),'MinStopCodons' (Number of stop codons in the selected frame), 'Frame' (putative frame),'Strand' (putative strand)
}
\description{
Translate a nulceotide sequence into all possible 6-frames and decide for the frame with the fewest number of stop codons
}

