% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDiPeptideFrequency.R
\name{GetDiPeptideFrequency}
\alias{GetDiPeptideFrequency}
\title{Get the frequency of di-peptides}
\usage{
GetDiPeptideFrequency(Seq, prob = F)
}
\arguments{
\item{Seq}{An AAStringSet object}

\item{prob}{Should relative frequncies be returned}
}
\value{
A data.frame object containing the frequencies (columns) for the submitted sequences (rows)
}
\description{
Get the frequency of di-peptides
}

