% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMonoPeptideFrequency.R
\name{GetMonoPeptideFrequency}
\alias{GetMonoPeptideFrequency}
\title{Get the Frequency of each amino acid}
\usage{
GetMonoPeptideFrequency(Seq, prob = F)
}
\arguments{
\item{Seq}{An AAStringSet object}

\item{prob}{Should relative frequncies be returned}
}
\value{
A data.frame object containing the frequencies (columns) for the submitted sequences (rows)
}
\description{
Get the Frequency of each amino acid
}
\author{
Carlus Deneke
}

