% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOligonucleotideFrequency.R
\name{GetOligonucleotideFrequency}
\alias{GetOligonucleotideFrequency}
\title{Get the ologonuleotide frequency}
\usage{
GetOligonucleotideFrequency(Seq, k = 1, prob = F, symmetric = F)
}
\arguments{
\item{Seq}{A DNAStringSet object, i.e. a set of DNA sequences}

\item{k}{The k-mer size}

\item{prob}{Return the relative frequency}

\item{symmetric}{Do you want to reduce number of kmers by combining forard kmers and their reverse-complements? (Default: FALSE)}
}
\value{
The ologonuleotide frequency for each sequence submitted for the choice of k
}
\description{
Obtain the frequency of oligonucleotides in a DNA sequence for different k-mer sizes
}
\author{
Carlus Deneke
}

