% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOSid.R
\name{GetOSid}
\alias{GetOSid}
\title{Extract the Organism ID (PRJNA accession)}
\usage{
GetOSid(x)
}
\arguments{
\item{A}{string containing an NCBI project accession (e.g. xxxPRJNA1752xxx)}
}
\value{
Returns a string with the extracted NCBI project accession
}
\description{
Extract the Organism ID (PRJNA accession)
}
\author{
Carlus Deneke
}
