% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFeatures.FromForest.R
\name{IdentifyFeatures.FromForest}
\alias{IdentifyFeatures.FromForest}
\title{Identify features from random forest object}
\usage{
IdentifyFeatures.FromForest(ForestObject, verbose = F)
}
\arguments{
\item{ForestObject}{An object of class 'ranger'}
}
\value{
A list object containing the relevant feature type information
}
\author{
Carlus Deneke
}

