% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Labels}
\alias{Labels}
\title{Pathogenicity labels}
\format{An object of class \code{logical} of length 3019.}
\usage{
Labels
}
\description{
Pathogenicity labels
}
\details{
The set of labels for 4286 bacterial strains (181 non-pathogens and 2838 pathogens)
TRUE means pathogenic and FALSE means non-pathogenic
The species' bioproject accession is obtained by names(Labels)
}
\references{
PaPrBag Paper
}
\keyword{datasets}

