% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.Symmetric.R
\name{Make.Symmetric}
\alias{Make.Symmetric}
\title{Make Features Symmetric by combining a dna string with its reverse complement}
\usage{
Make.Symmetric(Data)
}
\description{
Make Features Symmetric by combining a dna string with its reverse complement
}

