% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchOSlabel.R
\name{MatchOSlabel}
\alias{MatchOSlabel}
\title{Match the OSid (Bioproject id) to a vector of}
\usage{
MatchOSlabel(OSid, OSlabels)
}
\arguments{
\item{OSid}{A vector of query ids}

\item{OSlabels}{A vector with label information named with the corresponding ids}
}
\value{
Returns a vector of labels of the queried ids
}
\description{
Match the OSid (Bioproject id) to a vector of
}

