% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.ReadSet.fromFiles.R
\name{Predict.ReadSet.fromFiles}
\alias{Predict.ReadSet.fromFiles}
\title{Predict ReadSet}
\usage{
Predict.ReadSet.fromFiles(Path2Forest, Path2ReadFiles, saveLocation,
  OutputFilename = "Output", Return.Predictions = F, Save.AsTSV = F,
  verbose = T, ...)
}
\arguments{
\item{Path2Forest}{A path to a randomForest object of class ranger in .rds format}

\item{Path2ReadFiles}{A path to a fasta file containing the read files}

\item{saveLocation}{A path to the direcory where the output should be written to}

\item{OutputFilename}{The name of the outputfile (without file extension) Default = "Output")}

\item{Return.Predictions}{Should the predictions be returned}

\item{Save.AsTSV}{Should the output saved in tsv (tab separated format) instead of rds format ?}

\item{...}{Further arguments passed to function 'Predict.ReadSet'}
}
\value{
Either the predictions (if Return.Predictions == T ) or TRUE/FALSE if Outputfile exists
#' @seealso \link[ranger]{predict.ranger}
}
\description{
This function is a wrapper for function 'Predict.ReadSet'
}
\seealso{
\link{Predict.ReadSet}
}
\author{
Carlus Deneke
}
