% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.ReadSet.R
\name{Predict.ReadSet}
\alias{Predict.ReadSet}
\title{Make a prediction from a set of reads}
\usage{
Predict.ReadSet(ForestObject, ReadObject, Feature.Configuration = NULL,
  verbose = F, num.threads = NULL, ...)
}
\arguments{
\item{ForestObject}{A random forest object of class 'ranger'}

\item{ReadObject}{A DNAStringSet object containing the read sequences}

\item{Feature.Configuration}{A list containing all feature options. If set to NULL then features are extracted from forest object. See 'details' for more information. (Default = NULL)}

\item{verbose}{Print verbose output}

\item{num.threads}{Number of threads used for ranger prediction. Default is number of CPUs available.}
}
\value{
A prediction object of class 'ranger.prediction'
}
\details{
If the user supplies the Feature.Configuration it should contain the following list objects
\itemize{
 \item{"parameter 1"}{Stuff1}
 \item{"parameter 2"}{Stuff2}
 }
}
\seealso{
\link[ranger]{predict.ranger}

\link{Predict.ReadSet.fromFiles}
}
\author{
Carlus Deneke
}
