% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prosite_Top20}
\alias{prosite_Top20}
\title{Prosite patterns}
\format{An object of class \code{character} of length 20.}
\usage{
prosite_Top20
}
\description{
Prosite patterns
}
\keyword{datasets}

