% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ReadData}
\alias{ReadData}
\title{Example read data}
\format{An object of class \code{DNAStringSet} of length 21.}
\usage{
ReadData
}
\description{
Example read data
}
\details{
21 read sequences of length 125 in "DNAStringSet" format
}
\keyword{datasets}
