% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.Training.R
\name{Run.Training}
\alias{Run.Training}
\title{The workhorse function for the training workflow}
\usage{
Run.Training(Path2FeatureFile, Path2LabelFile = NULL,
  SelectedFeatures = NULL, savePath = NULL, ReturnForest = F,
  verbose = T, min.node.size = 1, num.trees = 100, mtry = NULL,
  importance = "impurity", num.threads = NULL, ...)
}
\arguments{
\item{Path2FeatureFile}{Path to feature file in rds format. Rows are independent data, columns are features. May also contain the label information  (no default)}

\item{Path2LabelFile}{Either a path to a file containing the label data or (default: NULL)}

\item{SelectedFeatures}{Either a vector of column numbers, or a vector of feature group names (default: NULL)}

\item{savePath}{Path to where the forest object "randomForest.rds" is written to. If NULL, no output is written(default: NULL)}

\item{ReturnForest}{Whether the forest object should be returned or not (default: F)}

\item{verbose}{(default: T)}

\item{min.node.size}{Option for ranger: Minimal node size. (default: 1)}

\item{num.trees}{Option for ranger: Number of trees. (default: 100)}

\item{mtry}{Option for ranger: Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables.}

\item{importance}{Option for ranger: Variable importance mode, one of 'none', 'impurity', 'permutation'. (default: 'impurity')}

\item{num.threads}{Option for ranger: Number of threads. Default is number of CPUs available.}

\item{...}{Additional parameters passed to function ranger, see \code{\link{ranger::ranger}}}
}
\value{
Returns either TRUE if forest has been saved successfully or returns the forest object itself
}
\description{
The workhorse function for the training workflow
}
\details{
Trains a random forest classification algorithm on the Feature file in Trainingfolder using the labels from LabelFile or LabelType. Saves the output (the random forest object) in a new subfolder of TrainingData
}
\examples{
\dontrun{
Run.Training (Path2FeatureFile = Path2FeatureFile, Path2LabelFile = Path2LabelFile, SelectedFeatures = NULL, savePath = savePath, ReturnForest = F, verbose = T, min.node.size = 1, num.trees = 100, mtry = NULL, importance = 'impurity', num.threads = NULL)
}
}

