% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectFeatureSubset.R
\name{SelectFeatureSubset}
\alias{SelectFeatureSubset}
\title{Function to select a subset of features from the feature table}
\usage{
SelectFeatureSubset(Features, SelectedFeatures)
}
\arguments{
\item{Features}{The original feature table as data.frame}

\item{SelectedFeatures}{Either a numerical vector containing columns to be selected or a vector of character strings to select groups of features}
}
\value{
The subsetted feature table as data.frame
}
\description{
Function to select a subset of features from the feature table
}
\seealso{
Other TrainingFunctions: \code{\link{Create.TrainingDataSet}},
  \code{\link{Run.Training}}
}
\author{
Carlus Deneke
}
