% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Standard.configuration}
\alias{Standard.configuration}
\title{Standard configuration for feature extraction}
\format{An object of class \code{list} of length 20.}
\usage{
Standard.configuration
}
\description{
Standard configuration for feature extraction
}
\details{
Choice of parameters:
NT_kmax = 4,
SymmetricFeatures = T,
Do.NTMotifs = T,
NTMotifs = NULL,
AllowedMismatches = 1,
Do.SpacedWords = T,
k.spaced = 4,
l.spaced = 6,
Do.PeptideFeatures = T,
Do.MonoPep = T,
Do.DiPep = T,
Do.AAprops = T,
Do.AAindex = T,
AAindex_Selection = NULL,
Do.UCO = T,
Do.PepPatterns = F,
Patterns = NULL,
SearchPatternsSixFrame = T,
AllowedPeptideMismatches = 0,
AggregatePatterns = F
}
\keyword{datasets}
