% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Standard.configuration_peptide}
\alias{Standard.configuration_peptide}
\title{Standard configuration for feature extraction including peptide features}
\format{An object of class \code{list} of length 22.}
\usage{
Standard.configuration_peptide
}
\description{
Standard configuration for feature extraction including peptide features
}
\details{
NT_kmax = 4
Symmetric = TRUE
Do.NTMotifs = FALSE
NTMotifs = NULL
AllowedMismatches = 1
bothStrands = TRUE
Do.SpacedWords = TRUE
k.spaced = 4
l.spaced = 6
SingleSpacerPattern = 0
combineSpacerPatterns = TRUE
Do.PeptideFeatures = TRUE
Do.MonoPep = TRUE
Do.DiPep = FALSE
Do.AAprops = FALSE
Do.AAindex = TRUE
AAindex_Selection = c("BROC820101","BUNA790103","CHOP780207","FAUJ880105","FINA910103","GEIM800103","GEIM800105","ISOY800107","KHAG800101","LEWP710101","MAXF760103","OOBM850104","PALJ810111","PRAM820103","QIAN880102","QIAN880114","QIAN880123","QIAN880137","RACS770103","RACS820103","RICJ880101","RICJ880117","ROBB760107","TANS770106","TANS770108","VASM830101","WERD780104","AURR980102","AURR980116","AURR980118","FUKS010109","SUYM030101")
Do.UCO = TRUE
Do.PepPatterns = FALSE
SearchPatternsSixFrame = FALSE
AllowedPeptideMismatches = 0
AggregatePatterns = FALSE
}
\keyword{datasets}
