% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainingData}
\alias{trainingData}
\title{Example training data
Provides a small set of training Data (combined label and feature table). First column name is "Labels", other column names indicate feature type}
\format{An object of class \code{data.frame} with 200 rows and 349 columns.}
\usage{
trainingData
}
\description{
Example training data
Provides a small set of training Data (combined label and feature table). First column name is "Labels", other column names indicate feature type
}
\keyword{datasets}

