% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateFeatures.R
\name{UpdateFeatures}
\alias{UpdateFeatures}
\title{Update all features in a specified folder}
\usage{
UpdateFeatures(Path2File, Cores = 1, Backup = T, ...)
}
\arguments{
\item{Path2File}{A path to the location of the read files}

\item{Cores}{Multicore functionality for linux systems.}

\item{Backup}{Do you want to backup old features}

\item{AAindex_Selection}{Parameter for Features}
}
\value{
Returns TRUE if completed
}
\description{
Update all features in a specified folder. Save features in subfolder "Features". Backup old features if desired.
}

