% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count.SpacedWords.R
\name{Count.SpacedWords}
\alias{Count.SpacedWords}
\title{Count the occurrence of all spaced words}
\usage{
Count.SpacedWords(Reads, k, l, combinePatterns = T, SinglePattern = 0,
  prob = T, symmetric = T)
}
\arguments{
\item{k}{The length of the contingous words, i.e. the kmer size}

\item{l}{The length of the spaced words}
}
\value{
A matrix containing the combined counts for all possible spacer-patterns
}
\description{
Count the occurrence of all spaced words
}

