% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.TrainingDataSet.R
\name{Create.TrainingDataSet}
\alias{Create.TrainingDataSet}
\title{Create a new training data set}
\usage{
Create.TrainingDataSet(Path2Files = NULL, OSlabels = NULL,
  savePath = NULL, CompressOption = T)
}
\arguments{
\item{Path2Files}{A complete path to the read folder (i.e. it contains a subfolder 'Features')}

\item{OSlabels}{A vector containing either 'HP' or 'NP' together with the name attributes pointing to the Organism identifier (Bioproject ID)}

\item{savePath}{The name of the newly created training data set folder}

\item{CompressOption}{Do you want to compress the saved files (takes time but saves disk space)?}
}
\value{
Returns True if completed. Feature and label files are saved in specified folder (see details)
#' @examples
\dontrun{
Create.TrainingDataSet (Path2Files = NULL, OSlabels = NULL, savePath = NULL,CompressOption = T)
}
}
\description{
Create a new training data set from a set of feature files. It concatenates all feature file into a data.table object. Read labels are assigned from the corresponding organism labels. Data are saved in specified folders.
}
\details{
Saves 3 files in specified folder "TrainingFolder".
FeatureTable.rds containing the features in a data.frame object
FeatureRowDescription.rds containing the read description and allows identification of Organism, Chromosome and read
ReadLabel_OS.rds containing a label for each read based on the organisms label
}

