% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.SymmetricPartners.R
\name{Match.SymmetricPartners}
\alias{Match.SymmetricPartners}
\title{Convert vector of sequences to its reverse complement and match symmetric partners}
\usage{
Match.SymmetricPartners(seq)
}
\description{
Convert vector of sequences to its reverse complement and match symmetric partners
}

