% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prosite_Top20}
\alias{prosite_Top20}
\title{Prosite patterns}
\format{An object of class \code{character} of length 20.}
\source{
http://prosite.expasy.org/prosite_ref.html
}
\usage{
prosite_Top20
}
\description{
Prosite patterns
}
\details{
20 frequent prosite motifs
}
\references{
Sigrist CJA, de Castro E, Cerutti L, Cuche BA, Hulo N, Bridge A, Bougueleret L, Xenarios I.
New and continuing developments at PROSITE
Nucleic Acids Res. 2012; doi: 10.1093/nar/gks1067
}
\keyword{datasets}

